Fix the panic() that happens when we try to deserialize JSONLines
Frame, which has newly acquired the ability to _deserialize_ itself properly, was embedded in JSONLine. This meant that the new MarshalJSON() function, on _Frame_, would actually take over and be the MarshalJSON for JSONLine, which was NOT the intention.

This would have been caught with a very simple test, and also _was_ caught with our live testing, which I failed to run. So as penance I've also added in a unit test.